修改 修复分组狂点导致崩溃问题 小夜灯页面调整

+++ app/src/main/java/com/dadoutek/uled/group/LightGroupingActivity.kt
@@ -43,36 +43,53 @@ class LightGroupingActivity : TelinkBaseActivity(), EventListener<String> {
     private var gpAdress: Int = 0
 
     private var listView: GridView? = null
+    private var isStartGroup = false
     private val itemClickListener = OnItemClickListener { parent, view, position, id ->
         val group = adapter!!.getItem(position)
         if (group != null) {
             if (TelinkApplication.getInstance().connectDevice == null) {
                 ToastUtils.showLong(R.string.group_fail)
             } else {
-                showLoadingDialog(getString(R.string.grouping))
-                object : Thread({
-                    val sceneIds = getRelatedSceneIds(group.meshAddr)
-                    deletePreGroup(light!!.meshAddr)
-                    Thread.sleep(100)
-                    deleteAllSceneByLightAddr(light!!.meshAddr)
-                    Thread.sleep(100)
-                    allocDeviceGroup(group)
-                    Thread.sleep(100)
-                    for (sceneId in sceneIds) {
-                        val action = DBUtils.getActionBySceneId(sceneId, group.meshAddr)
-                        if (action != null) {
-                            Commander.addScene(sceneId, light!!.meshAddr, action.color)
+                if(!isStartGroup){
+                    isStartGroup=true
+                    showLoadingDialog(getString(R.string.grouping))
+                    object : Thread({
+                        val sceneIds = getRelatedSceneIds(group.meshAddr)
+                        for(i in 0..1){
+                            deletePreGroup(light!!.meshAddr)
+                            Thread.sleep(100)
                         }
-                    }
-                    DBUtils.updateLight(light!!)
-                    runOnUiThread {
-                        hideLoadingDialog()
-                        ActivityUtils.finishActivity(RGBSettingActivity::class.java)
-                        finish()
-                    }
-                }) {
 
-                }.start()
+                        for(i in 0..1){
+                            deleteAllSceneByLightAddr(light!!.meshAddr)
+                            Thread.sleep(100)
+                        }
+
+                        for(i in 0..1){
+                            allocDeviceGroup(group)
+                            Thread.sleep(100)
+                        }
+
+                        for (sceneId in sceneIds) {
+                            val action = DBUtils.getActionBySceneId(sceneId, group.meshAddr)
+                            if (action != null) {
+                                for(i in 0..1){
+                                    Commander.addScene(sceneId, light!!.meshAddr, action.color)
+                                    Thread.sleep(100)
+                                }
+
+                            }
+                        }
+                        DBUtils.updateLight(light!!)
+                        runOnUiThread {
+                            hideLoadingDialog()
+                            ActivityUtils.finishActivity(RGBSettingActivity::class.java)
+                            finish()
+                        }
+                    }) {
+
+                    }.start()
+                }
             }
         } else {
             //                Toast.makeText(mApplication, "", Toast.LENGTH_SHORT).show();
